requirements.txt
pymongo
pinecone-client
mysql-connector-python
psycopg2-binary
pymongo[srv]
dnspython
pandas
numpy
ensure


-e .